user-preference-service/requirements.txt
fastapi
uvicorn
python-dotenv
mysql-connector-python
aio-pika>=7.0.0
aiormq>=6.0.0